Split integration tests
Don't run tests which use real GCS.
This way, "go test" always works, even without a GCS bucket.

Integration tests, which require a GCP account can be run separately
with:

    go test -tags integration
