+++ mutex_test.go
@@ -5,22 +5,8 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"reflect"
-	"sync"
 	"testing"
 	"time"
-)
-
-const (
-	project = "marc-general"
-	bucket  = "gcslock"
-	object  = "lock"
-)
-
-var (
-	limit = 10
-
-	lockHolderMu sync.Mutex
-	lockHolder   = -1
 )
 
 func TestLock(t *testing.T) {
@@ -45,7 +31,7 @@ func TestLock(t *testing.T) {
 	defer storage.Close()
 	storageLockURL = storage.URL
 
-	m, err := New(nil, project, bucket, object)
+	m, err := New(nil, "stub", "gcslock", "lock")
 	if err != nil {
 		t.Fatal("unable to allocate a gcslock.mutex object")
 	}
@@ -77,7 +63,7 @@ func TestLockRetry(t *testing.T) {
 	defer storage.Close()
 	storageLockURL = storage.URL
 
-	m, err := New(nil, project, bucket, object)
+	m, err := New(nil, "stub", "gcslock", "lock")
 	if err != nil {
 		t.Fatal("unable to allocate a gcslock.mutex object")
 	}
@@ -112,7 +98,7 @@ func TestUnlock(t *testing.T) {
 	defer storage.Close()
 	storageUnlockURL = storage.URL
 
-	m, err := New(nil, project, bucket, object)
+	m, err := New(nil, "stub", "gcslock", "lock")
 	if err != nil {
 		t.Fatal("unable to allocate a gcslock.mutex object")
 	}
@@ -146,7 +132,7 @@ func TestUnlockRetry(t *testing.T) {
 	defer storage.Close()
 	storageUnlockURL = storage.URL
 
-	m, err := New(nil, project, bucket, object)
+	m, err := New(nil, "stub", "gcslock", "lock")
 	if err != nil {
 		t.Fatal("unable to allocate a gcslock.mutex object")
 	}
@@ -163,42 +149,6 @@ func TestUnlockRetry(t *testing.T) {
 	}
 	if retryCount < retryLimit {
 		t.Errorf("retryCount = %d; want %d", retryCount, retryLimit)
-	}
-}
-
-func locker(done chan struct{}, t *testing.T, i int, m sync.Locker) {
-	m.Lock()
-	lockHolderMu.Lock()
-	if lockHolder != -1 {
-		t.Errorf("%d trying to lock, but already held by %d",
-			i, lockHolder)
-	}
-	lockHolder = i
-	lockHolderMu.Unlock()
-	t.Logf("locked by %d", i)
-	time.Sleep(5 * time.Millisecond)
-	lockHolderMu.Lock()
-	lockHolder = -1
-	lockHolderMu.Unlock()
-	m.Unlock()
-	done <- struct{}{}
-}
-
-func TestParallel(t *testing.T) {
-	storageLockURL = defaultStorageLockURL
-	storageUnlockURL = defaultStorageUnlockURL
-	m, err := New(nil, project, bucket, object)
-	if err != nil {
-		t.Fatal("unable to allocate a gcslock.mutex object")
-	}
-	done := make(chan struct{}, 1)
-	total := 0
-	for i := 0; i < limit; i++ {
-		total++
-		go locker(done, t, i, m)
-	}
-	for ; total > 0; total-- {
-		<-done
 	}
 }
